Add files via upload
Journal part
The first thing I did was to download the dataset called ‘Indicadores de pobreza, pobreza por ingresos, rezago social y gini a nivel municipal,1990, 2000, 2005 y 2010’. The target I chose was ‘N_plb Personas con ingreso inferior a la línea de bienestar’, which was providing the numerical value. I was solving a classification problem, where it is necessary to have categories, therefore, I needed to transform my target from numerical to categorical. The question I was solving was to classify if the different populations depending on if most of the half of the population was over the wellness income line. With that, I was able to determine two categories: ‘1’ if more than the half of the population would be under the defined wellness income line and ‘0’ if less than the half of the population was on that situation.
Once I defined my approach, I started to make the necessary modifications to the dataset. Because I was going to work with classification algorithms, and non-numerical data can affect the performance of it, I decided to drop the non-numerical columns that I identified, also, the ones that only were mentioning the number of the municipalities, because my problem was only focused on working with the population regardless the origin of them, I was not going to take into account the name of the place or the number of the place, therefore, I decided to drop those columns too.
Also, it was necessary to fill the empty cells on the dataset, this because the missing data can lead to incorrect results, and in order to have the best results possible, I decided to input the missing values with the mean of the same column, in order to avoid the reduction of my data size. Also, it would make it compatible with the knn and perceptron algorithms applied to train the model.
For the conversion of my numerical target into a categorical one, I decided to create a new column  called ‘Conversion of target’ with the labels 1 and 0, as I explained in a previous paragraph, that would be the two categories my model will be trained for. The whole process consisted on: First, it was calculated the percentage that represented the people under the wellness income line for each population taking into account the total population and the numerical values from my original target; once that was calculated, it was assigned a value of 1 to the ones where the percentage of was more than 50% and ‘0’ when tit was less than 50% of the population. 
Once that new feature was created, I used it as the target for implementing the algorithms. It was already the last column of my dataset, so I did not need to make any modification related to that part.
The main challenge of this project, for me, was to try to think about how the best way for me was to prepare my dataset, and how I would transform my target into a variable that could be useful for applying both of my algorithms. Also, to understand how the different algorithms work, how they are conformed, their steps and how the hyperparameters could make the result of the estimations change. . Investigating the role of hyperparameters was an eye-opener, as even slight tweaks could drastically alter the results. The choice of these settings was a crucial decision in the project's success.
In the field of Robotics, this can be applied to the artificial vision area,  and the automation field, where tasks related with classification of different products, for example, in robotics, robots often encounter machine learning tasks such as object classification, autonomous navigation, and sensor-based decision-making. Both K-Nearest Neighbors (KNN) and the Perceptron are supervised learning algorithms that can be used for classification tasks, which are crucial in robotics.